refactor: rename .veryfrontrc to veryfront.json and consolidate init/new commands
- Rename .veryfrontrc → veryfront.json across all CLI commands and help text
- Remove `new` command, absorb --deploy flag into `init`
- Move reserve-slug.ts to cli/shared/ for reuse by push, demo, and init
- Update demo command to use initCommand instead of newCommand
- Delete new/command.ts, handler.ts, command-help.ts, tui.ts, menu.ts and tests
