+++ cli/commands/new/index.ts
@@ -1,9 +1,6 @@
 /**
- * New command - Lightning-fast project creation
+ * New command utilities - re-exports from shared
  */
 
-export { NewArgsSchema, newCommand, parseNewArgs } from "./command.ts";
-export type { NewOptions } from "./command.ts";
-export { handleNewCommand } from "./handler.ts";
 export { scaffoldProjectFast } from "./fast-scaffold.ts";
-export { reserveProjectSlug } from "./reserve-slug.ts";
+export { reserveProjectSlug } from "#cli/shared/reserve-slug";
