--- cli/commands/new/tui.test.ts
@@ -1,94 +1,0 @@
-/**
- * Tests for new-tui.ts (Charm-style wizard)
- */
-
-import { assertEquals, assertExists } from "#veryfront/testing/assert.ts";
-import { describe, it } from "#veryfront/testing/bdd.ts";
-import type { NewTuiResult } from "./tui.ts";
-
-describe("new-tui", () => {
-  it("exports runNewTui function", async () => {
-    const mod = await import("./tui.ts");
-    assertExists(mod.runNewTui);
-    assertEquals(typeof mod.runNewTui, "function");
-  });
-
-  it("works with ai template", () => {
-    const result: NewTuiResult = { template: "chat", integrations: [], cancelled: false };
-    assertEquals(result, { template: "chat", integrations: [], cancelled: false });
-  });
-
-  it("works with integrations", () => {
-    const result: NewTuiResult = {
-      template: "rag",
-      integrations: ["gmail", "slack", "github"],
-      cancelled: false,
-    };
-    assertEquals(result.template, "rag");
-    assertEquals(result.integrations, ["gmail", "slack", "github"]);
-    assertEquals(result.cancelled, false);
-  });
-
-  it("works when cancelled", () => {
-    const result: NewTuiResult = { template: "chat", integrations: [], cancelled: true };
-    assertEquals(result.cancelled, true);
-  });
-
-  it("supports all template types", () => {
-    const templates: NewTuiResult["template"][] = [
-      "chat",
-      "rag",
-      "multi-agent",
-      "workflow",
-      "coding-agent",
-      "saas",
-      "minimal",
-    ];
-    assertEquals(templates, [
-      "chat",
-      "rag",
-      "multi-agent",
-      "workflow",
-      "coding-agent",
-      "saas",
-      "minimal",
-    ]);
-  });
-
-  it("supports all integration types", () => {
-    const integrations: NewTuiResult["integrations"] = [
-      "gmail",
-      "slack",
-      "notion",
-      "github",
-      "calendar",
-      "drive",
-      "jira",
-      "linear",
-    ];
-    assertEquals(integrations, [
-      "gmail",
-      "slack",
-      "notion",
-      "github",
-      "calendar",
-      "drive",
-      "jira",
-      "linear",
-    ]);
-  });
-});
-
-// Interactive tests require manual testing:
-// deno task cli new my-test-project
-//
-// Test checklist:
-// [ ] Template selection with arrow keys (h/j/k/l and arrows)
-// [ ] Template selection with mouse click
-// [ ] Integration selection with arrow keys
-// [ ] Integration toggle with space
-// [ ] Toggle all with 'a'
-// [ ] Mouse click on integrations
-// [ ] Enter to confirm
-// [ ] Ctrl+C to cancel
-// [ ] Alt screen mode (clean exit)
