+++ cli/shared/slug.ts
@@ -1,0 +1,12 @@
+/**
+ * Slug utilities for project naming
+ *
+ * @module cli/shared/slug
+ */
+
+export function randomSuffix(len = 6): string {
+  const chars = "abcdefghijklmnopqrstuvwxyz0123456789";
+  return Array.from({ length: len }, () => chars[Math.floor(Math.random() * chars.length)]).join(
+    "",
+  );
+}
